+++ infrastructure/terraform/lambda.tf
@@ -1,3 +1,9 @@
+# CloudWatch Log Group with retention period
+resource "aws_cloudwatch_log_group" "task_tracking_list" {
+  name              = "/aws/lambda/text-agent-task-tracking-list"
+  retention_in_days = 14
+}
+
 resource "aws_lambda_function" "task_tracking_list" {
   function_name = "text-agent-task-tracking-list"
   role          = aws_iam_role.lambda_exec.arn
@@ -13,6 +19,7 @@ resource "aws_lambda_function" "task_tracking_list" {
 
   depends_on = [
     aws_iam_role_policy.lambda_exec_policy,
+    aws_cloudwatch_log_group.task_tracking_list,
   ]
 }
