+++ infrastructure/terraform/task_tracking_action_group.tf
@@ -1,22 +1,22 @@
 # Create an action group for the Bedrock agent to enable task tracking tools
-resource "aws_bedrockagent_agent_action_group" "task_tracking" {
+resource "aws_bedrockagent_agent_action_group" "task_tracking_list" {
   # Reference the agent ID and set the version to DRAFT
   agent_id      = aws_bedrockagent_agent.text_agent.agent_id
   agent_version = "DRAFT"
 
   # Define the action group
-  action_group_name = "TaskTracking"
+  action_group_name = "TaskTrackingList"
 
   # Inline schema that defines the available actions
   function_schema {
     member_functions {
       functions {
-        name        = "task_tracking"
-        description = "Task Tracking"
+        name        = "task_tracking_list"
+        description = "Task Tracking List"
         parameters {
-          map_block_key = "action"
-          type          = "string"
-          description   = "The action to perform"
+          map_block_key = "phone_numbers"
+          type          = "array"
+          description   = "The phone numbers to list tasks for"
           required      = true
         }
       }
